requirements.txt
PyYAML==6.0
requests==2.25.1
requests_oauthlib==1.3.1
tqdm==4.64.1
tweepy==3.10.0
